SA: Fix div/0 in case of 0-sized device.
Simple and stupid.

+++ sa.c
@@ -68,7 +68,7 @@ static int lfsr(unsigned int v, unsigned int bits)
 {
 	unsigned int bit;
 
-	if (v >= (1 << bits)) {
+	if (v >= ((unsigned int) 1 << bits)) {
 		fprintf(stderr, "sa: internal error\n");
 		exit(-EINVAL);
 	}
